Add docker image that runs the UI using node
This adds a Dockerfile that will build an image with the current
composer-UI code and run it using the 'node run start' command.

This also changes how the composer_api_host variable is set. In order
to support changing it at runtime instead of image build time it has
been moved into ./public/js/config.js and is loaded as part of the
index.ejs code.

The entrypoint.sh script checks for a API_URL environmental variable and
passes it on to config.js, or falls back to using http://localhost:4000

You can build the image with:

docker build -t bcl/composer .

And you can run it, passing an API URL, with:

docker run -it --rm -e API_URL=http://192.168.102.4:4444 bcl/composer