Fix the recipe list display.
Getting to the recipe object the hard way was a side effect of sloppy
example data and mine own poor understanding of what a javascript
promise actually closes over. As a side-effect, when running under json
recipe, the recipe list ended up with a ghost recipe from the extra "id"
attribute. Bust the ghost.