When running the API behind a proxy it needs to use relative URLs
Add a switch to make createUrl just return the url, not add the scheme,
host or port.

Split the API_URL into component pieces for new config.js, ALL parts
must be present when setting it.

This also displays the modified config.js when starting.

Startup of nginx has been moved to CMD in the Dockerfile so that you can
start up the container without overriding entrypoint. Just pass
/usr/bin/bash to run bash instead of nginx