updated component list item data to indicate user-selection
This update helps track whether the user is looking at a package that
is in the recipe and the user selected it or did not select it (i.e. it's a
dependency). This lets the UI display different actions (e.g. in the component
details view) for dependencies vs packages selected by the user.