updating UI to either show data or "---"
Areas in the UI where data might be helpful but isn't
available yet were updated to display "---". This includes
replacing data for arch with "---" since it's questionable if
this is needed.

Areas in the UI where data is available but not displayed
correctly were also updated.

This update also addresses some lint issues