fixing a bug related to pagination
When preventDefault and stopPropagation were
added to the pagination, they prevented the default
behavior of entering a new value in the input field.
I moved these to only the cases where the user is
hitting the enter key when the input has focus or
using the pagination buttons, but not when hitting
any other key (e.g. like "9" to navigate to page 9).