Collect coverage from all end-to-end tests and save it under /tmp
after the application code has been instrumented we use
nightmare.evaluate() to fetch the data from browser scope back to
node scope and save it in a json file under /tmp. This file is
later used for reports.

Each nightmare sequence must not call .end() and done() because
this is done in coverage.js as the last step of the test!

NOTE: we're using eval() on the contents of utils/coverage.js
because all my attempts to successfully execute this piece of code
via a helper module failed! eval() in this scope is equivalent to
inserting the contents of utils/coverage.js into the test!