Replace Overview page with blank slate pattern on Recipes
The Overview page is very empty and is mostly just tailored for
usability testing with pre-existing recipes. For real deployments with
no pre-existing recipes it's a more common workflow to put the
instructions as a "blank slate pattern" on the actual recipes page.

This works towards avoiding the navigation menu completely (together
with dropping the Users page).

Rename home.index.spec.js unit test to index.norecipes.spec.js and
verify that blank slate is generated with zero recipes.