Hide header bar in Cockpit
It is confusing to have *two* header bars with user menus. None of the
current entries actually do anything, but if/once they do they need to
be integrated into the Cockpit UI in a different way.

Drop the "layout-pf-fixed" class for that. In Cockpit mode it otherwise
causes a 60px top margin where the header bar was, and it doesn't change
the appearance in standalone mode.

Closes #150