Don't ignore failed serviceCheck in tests
The integration test's `serviceCheck()` previously caught errors (if the
checked services like bdcs and welder-web are not running) and just
called the `done()` callback, but that does not care about the returned
value - in particular if that is an exception. This causes unhelpful
test failure output and makes the entire `serviceCheck()` call rather
useless.

Instead, directly return the two involved promises instead of using a
`done` callback, to let jest deal with it properly (see
http://facebook.github.io/jest/docs/en/setup-teardown.html)

Drop the unused `webCheck()` and `apiCheck()` methods.