Version release tarballs
For doing releases we need to produce tarballs with versions in the
name.

Generate version number from the latest git tag; if the topmost commit
is tagged (at release time), just use that; if there are commits after
that, append a ".x" suffix to indicate that's a development snapshot.
Exclude jenkins-related temporary tags that our CI produces.

Use the version number for `make dist-gzip` and adjust the spec file
accordingly.

Also add missing `%changelog` tag to spec file.