A little changes to force three cases fail, then catch the failed workflow.
This's a test commit to test new feature added in this PR. The failed workflow
will be triggered by three failed cases. The error message will be at Travis
console, and screenshot will be on S3(NOT WORK FOR PR).

NOTE: The uploading to S3 will not work on PR due to Travis PR security restriction.
https://docs.travis-ci.com/user/pull-requests/#Pull-Requests-and-Security-Restrictions
But this part of result can be found from https://travis-ci.org/henrywang/welder-web/builds.
There'll not have this kind of issue, because it's not a PR, but a pushed branch.