Leave only tar image type for testing Create Image notifications
these are backend dependent and may actually take a lot of time
if/when the backend decides to really build a compose. So test
only with the tar compose type b/c it should be relatively fast.

There's a JIRA ticket to add more testing of composes so we can
do a sanity compose test with all supported types there.