Add translated messages.
Before the build and publish targets, pull available translations from
Zanata and convert them to a JSON file usable by react-intl. As part of
the build target in run.js, generate a module to load the necessary
locale data for any language that has translations. In main.js, check if
the language set in navigator.language has translations available, and
if so set them in the IntlProvider.

For the future, we'll probably want a way to override the language set
in navigator.language, we may want to store the translations in Redux,
and there's probably a more elegant way of dealing with the data than
all of those modules generated in run.js. But for now this works.