Create cockpit translation modules.
For each translation file available, create a module for cockpit
containing the translations. We only need these modules to initialize
the locale data within cockpit and to provide translations for the
manifest label, but just include the entire .po file anyway so that we
don't need to do anything special.

po2json and po.empty.js are copied from the cockpit sources, and handle
converting .po files into modules in the format that cockpit requires.
Despite the name, the data format is very much not JSON, so beware if
attempting to modify the scripts. po.empty.js is modified from the
cockpit source to include an eslint-disable comment.

In main.js, check if we are running within cockpit, and if so use
cockpit's language information to select the messages to use with
react-intl.