Add a script for testing translations.
`npm run translations:test` will generate (bad) Pig Latin translations
of all translatable strings and save them as "Volapük". Volapük is used
because it's supported by react-intl for some reason, and because there
the number of speakers is in the double digits so it's highly unlikely
to ever be actually used.