core: propagate errors from cockpitFetch()
Errors from cockpit.http.request() were only printed to the console.
Properly reject() the promise with a structured error, so that the UI
doesn't get stuck on an infinite spinner. This was most noticable when
logging in as a user without privileges.

Also update all call sites of cockpitFetch() to print the new error,
which is now an object.

Fixes #293