Remove unused Layout components
Removes Header and logic for displaying Header based
on whether this is in the context of Cockpit or not. Header
and this logic can be added back if this ever becomes a
stand-alone app.
Updates styles for toast notifications to be visually
consistent with PatternFly. PF styles are based on
the presence of a header. Updated styles are based on
this app loading in an iframe with the header outside the
iframe.