Add a po-push target to the Makefile.
Yay, symmetry!

Additionally, to make this target usable, for instance, in the cloud
with authentication data crammed into the environment, add a wrapper
script to translate $ZANATA_USER and $ZANATA_API_KEY into arguments to
zanata-js. If these environment variables are not set, zanata.ini will
be used like before.