Update Create Image modal to also commit unsaved changes
If a blueprint has unsaved changes, the Create button changes to
"Commit and Create" and will commit changes before staring
the image build process.
The function handleCommitBlueprint is also updated to accept
the blueprint as an argument to support this function in the context
of the Create Image modal, but ultimately, this function should be
refactored. For now, it is copied from the EditBlueprint page.