Changes en-dash to dash and adds spaces
These changes help with screen reader support. Screen
readers don't recognize the en-dash, which is the typographical
equivalent of "to", but screen readers will announce a
regular dash as "dash".
Additionally, if a dash is inserted between two numbers
without spaces, the individual digits are announced. However
if spaces surround the dash, then the numbers are announced
as whole numbers (e.g. "five zero" vs "fifty").