Add ability to stop builds that are waiting or running
Waiting builds are removed immediately. Running builds will
display a confirmation modal to the user first.
The Cancel api call doesn't seem to run as fast as the Delete
one, so there's an additional queue_status 'STOPPING' added
to the UI state for removing the Stop button while waiting on the
api call to finish.
Updates are needed to the existing pollComposeStatus
function to break the While loop if the polledCompose was
stopped by the user.