Temporarily support selenium images with and without -debug variants
https://github.com/cockpit-project/cockpit/pull/10191 will update
Cockpit's selenium image to have (only) the -debug variants of the
firefox and chromium containers. Temporarily support both, so that we
can run welder-web tests with both the current and that new image, and
never have to break tests.

This can be simplified to only using the -debug variants after landing
the above PR.