new test: Build mock images and verify download becomes enabled
this builds on previously existing tests.

Note: not trying to download the images b/c catching the
confirmation dialog is impossible (it is a system dialog). The other
approach is to configure the browser to download w/o confirmation
but we have no easy way of knowing when the download has completed!