Updates Create Image modal to not have a default image type
Instead of defaulting to the first image type in the select menu,
the select menu displays "Select one" by default. This is a disabled
option which means it can't be selected in the menu.
This menu is also indicated as required to prompt the user to make
a selection, and the submit button is disabled until an option is
selected.

Closes #374