test: Only install node if it isn't already available
Cockpit's rhel-7-6 image will soon pre-install node, as this is brittle
and time consuming when running for each test. As this code is not
reentrant, check if node is already installed before installing it.

With that we can update the cockpit image without breaking welder-web in
between. After that landed, this can go completely.