Remove inert 'Architecture' field in 'Create Image' dialog
Composer creates images in the architecture of the system it
is running on, and based on the source of the RPMs it has
access to (at least for the forseeable future).

This means this field in the Create Image dialog doesn't
do anything. And seems to be filled with placeholder text.