Fix "react/require-default-props" eslint error.
1. components/ListView/BlueprintContents.js
errorState is an object and default value is {}, so the condition
errorState !== undefined does not work. Update this condition to
(Object.keys(errorState).length to make sure it's not empty.
2. components/ListView/ListItemComponents.js
Fix two bugs here:
a. Compare two object by === does not work here, we care about content
not reference. So update it to JSON.stringify(olditem) !==
JSON.stringify(newitem)
b. the code uses [data-name=xxx] to locate page element, but no
this attribute in this component. So just add this attribute back
and rename it to [data-component=xxx] because "-name" is too common.