Fix some bugs which caused by eslint error fixing.
1. pages/blueprintEdit/index.js
The parent prop has two possible values, string and object.
2. components/Pagination/Pagination.js
The init state of pageValue should be 0 not "".
3. components/ListView/ComponentInputs.js
The key should be unique and not be array index. The solution is
use shortid lib to generate a unique short string as key. The
component name cannot be used as key because it's not unique here.