Refactor api call for inputs (i.e. "Available Components")
Prior to these changes, we would get the blueprint components
then get the inputs, then update the input component data
so that we can correctly show an Add or Remove action based
on whether the input is in the blueprint or not. The way this worked
was a little flimsy, and relied on a timeout after adding/removing
a component to update this set, which resulted in fetching the
list again and didn't really work. Instead, a selector is used, so that
once both lists are available or are updated, the selector compares
the inputs to the blueprint components to display them correctly.
As part of this change, some css was added to indicate which
components are selected vs dependencies using the
ComponentTypeIcon.
And the API call is refactored and uses saga to handle getting
the inputs and input component metadata.