Refactors how we get component details
Component details display in a Component Details panel after clicking
on a component and also in the expanded section of a component list item.
These updates use redux saga for managing api calls related to getting
metadata and getting a list of dependencies for a component.
NOTE: updates for handling component details in the Blueprints saga
are included in the next commit. They couldn't be decoupled from other
changes in that commit.