Refactor how blueprint components are updated and workspace
Prior to this change, the dependencies were updated using
get_project_deps, which returns a different set of dependencies
compared to get_blueprint_deps. Now when a component is added
or removed, the updated blueprint is posted to the workspace,
which then allows us to depsolve the blueprint, since that depsolves
the blueprint stored in the workspace.
Updating selected components is immediately reflected in the store,
while the get_blueprint_deps api call is made.
Selecting a different package version when one is available is
now supported (i.e. changes are now add/update/remove and not
just add/remove).
When there are changes in the workspace on first loading a blueprint,
the user can Undo or Discard Changes those changes, and still hit
Redo to get them back.