catch depsolve error and display error message
This pulls the message provided by the api when the
contents cannot be depsolves, and displays an error
state when this happens.
In the case where this happens as a result of an action
in the UI, the user also sees an Undo action in the error
message to revert to the previous state of the blueprint.
To make this change, the errorState has to be stored as
part of the blueprint history.