refactor Edit Description action
Updates Edit Description to result in a version bump
(previously, editing the description would not affect the
version value).
Refactors the existing Edit Description function and
moves it from BlueprintApi to sagas/blueprints.js.
Introduces a component based on the PF3 css for
inline edit, and includes a slight fix to the css.
Moves the input value from the redux store to the
component state to avoid lagging while typing.
Improves the keyboard interaction associated with
this action (text in input is automatically highlighted to
enable easier editing, Esc key will dismiss the input
and any changes made).