Include ability to add a user to the blueprint
User account creation allows for defining name, description,
password, ssh key and also enabling the user to be a server
admin (i.e. groups=["wheel"]).
User accounts can also be updated or removed.
User accounts that are created for a blueprint are listed on
the Details tab of a blueprint.