Addresses issue with duplicate keys when multiple arches
When the api call constants.get_blueprints_deps returns a list
of dependencies, if a package has more than one arch, then
each arch is a separate item in the array of dependencies.
Since the list item key is based on the name, this results in
issues when rendering the list of contents as they are updated
due to the key not being unique.