fixes issue with fetching contents for Export modal
Unlike the other pages, the Export modal on the Blueprints page
requires an api call to fetch the contents of the modal. This was
broken in #500 when the api calls were refactored. This update
moves the functionality from the modals saga to the blueprints
saga to reuse functions that handle flattening the list of
components after the api call.