Resolve issues with changes saved to the workspace
The Create Image modal should display a warning if changes are
saved to the workspace.
The example blueprints have some packages defined as "Modules"
which results in issues with how these are listed in the Pending
Changes modal if they are removed or updated without immediately
committing them (i.e. if they are retrieved from the workspace in a later
session).