Add link to show the anaconda.log from the compose build process
Especially when some image build failed it's usefull to be able to see
from the composer UI the anaconda logs.
This commit enables this functionality, by having a link for showing the
logs in each image.

The extra 'format' parameter was added to the requests method in
composer.js because the /compose/log API returns raw text, not json.

Closes #686