Fix warnings appearing in console when API calls failed
Currently we '.catch' the API errors when performing the API calls. This
results that the sagas will not fail when calling the API, and the code will
continue with executing the commands in the try block.
Right after the sagas will call the reducers but without the action payload
initialized and will result in invalid references causing warnings like
the following:

** uncaught at fetchModalManageSourcesContents TypeError: Cannot read property 'sources' of undefined

This commit removes the '.catch' from all api calls inside apiCalls.js
since there is fail handler in the sagas which call these APIs.

In the future we need to propagate some of these errors in visible UI
alerts, since we can't expect the users to read the JS console to find
out for errors.

Closes #680