Refactor Export Blueprint dialog to use Modal from patternfly-react
The part of the redux state reffering to the Export modal was removed,
since most of the state fields can be passed as properties to the
ExportBlueprint component.
The ones that need to be inside the global state where moved to the
blueprint object.

Note: The fetchBlueprintExportContents is called now only when we click on the
export button, and not when loading the blueprints page.