Add upload image functionality and wizard
The create image modal is replaced with a wizard that creates an image
and gives the user the option to upload the image to a cloud provider.
This wizard guides the user through selecting where to upload an image
and has the user enter the settings needed to upload the image. The
redux store and sagas are updated to handle the upload data.