Remove all functions which were supposed to keep track of the active tabs
There is not any benefit to keep the active tabs in the store.
These information should be kept in the state of the component that
creates the Tabs and not somewhere in the global state, but we are using
Tabs as uncontrolled component here, so we don't need to maintain the
state of the active tabs at all.