test: Fix "View Blueprint Page image type should be tar" case failed on RHEL-7-8
Fix failed case issue:
    1) View Blueprint Page image type should be tar:
    expected '' to equal 'tar'

The failed case is because the openssh-server dependencies on RHEL 7.8
have tar included. The selector 'strong=tar' will have two elements
matched, and the first matched element is the dependence one, not the
image type. But the tar is not one of dependencies of openssh-server on
Fedora or RHEL 8. So this issue cannot be found on those two test
scenarios.