Updates the blueprints saga to generate a list of components on depsolve failure
In the saga, instead of just returning the depsolve error, the list of components are
still generated from the blueprint data. Also, the depsolve error is added to the state
at the same time the list of components are added, to avoid displaying the empty
state message in BlueprintContents.

Also includes similar updates to the saga function for reloading the components after
adding/removing components to a blueprint.

BlueprintContents is also updated to show this list of components, along with the
error message. The next commit will replace the error message with an inline Alert
and also display warning icons on the Dependencies tab and in the list of selected
components for components that have issues with the specified version.

Fixes #718