Adds ability to update version selection for packages that cause error
If the specified version is not a version that can be selected, then the first option
(i.e. "*") will be selected by default when the user navigates to the component
details view for the package.

Since updating the component list can cause and resolve errors, the reducer
is updated to clear the previous errorState for the blueprint and also replace
it with a new errorState if one occurs as a result of updates. The update made
in /sagas/blueprints.js in this commit probably should have been made in the
first commit, but this wasn't change obvious until working on this flow.

An issue was also noticed when working on this flow in how we handled
updating components. Originally, component updates were always assumed to
be in the packages array. This commit includes a check to make sure that
package and module arrays are maintained as originally defined. Previously,
updating a component that was defined as a module in the blueprint would
add the component to the packages array, and leave the old version of the
component in the modules array.