tests: remove testing for default blueprints
With the migration from lorax-composer to osbuild-composer the default
blueprints have been removed. The tests that depended on having 3
default blueprints at initialization are updated. Also, the
blueprints page's test's loading function is updated to check for
the create blueprint button instead of the list of blueprints since
initially there is no list of blueprints. The existence of the
create blueprint button proves the page has loaded.