add image size input to create image upload modal
A user may want to manually specify an image size. They may now enter a
size into an input field. The size defaults to 2 GB which is also the
minimum size allowed. If the size is over 2000 GB an alert is shown
explaining that this image may not work on all systems.