test: Always explicitly install lorax-composer
In Fedora, golang-github-osbuild-composer Provides: lorax-composer, but
does not currently fully replace it (e. g. it doesn't provide
lorax-composer.service). Also, we want to be explicit about which
backend we test against. So always explicitly install lorax-composer for
now.

In the future, we can add another scenario that tests against osbuild,
but that needs slightly different setup.