components: add help text for aws
A user may be confused about what each field in the aws upload steps
means. Popover help text is added providing an explanation for each
field. General help text about uploading to aws is also added. The
structure of the upload and review steps of the CreateImageUpload wizard
have been restructured to enable the help text.