modals: implement create image upload wizard
A user may want to upload an image to a cloud service provider. The
create image modal has been updated to be a wizard. When creating an
ami a user has the option to upload it to aws. If they select to upload
it to aws they will be prompted for the needed upload arguments. The
create image modal has also been renamed to the create image upload
modal. Therefore, several files have been updated per the name changes.
Also, the tests relating to image creation or the create image button
are updated.