components: update image size help text
When a user wants to create an image the image size input should have
help text stating what the minimum image size is. The minimum image size
was only displaying when the user's input size was too small. The helper
text now always displays but is only displayed as a danger when the size
is too small.