components: improve aws upload review step
When a user chooses to upload to aws they can review the data they
entered in the final step of the CreateImageUpload wizard. The review
step is updated to improve how the fields are organized and presented.
Also, there is an additional aws popover containing help text in the
review page.