test: Fix preferences for new Firefox
Cockpit's CI moved from an old Firefox 69 to the current version and
nightly builds. This requires some additional user preferences which are
fixed in cockpit master, but not in a release yet.

So check out the corresponding SHA from cockpit's test/common. Once
Cockpit 221 gets released, it can go back to cloning the tag.

Closes #988