test: resize test vm root partition
The default test image created by the cockpit testing framework is too
small for some of our image builds. In these instances the vm will need
to be resized. When the vm is being setup vm.install is called.
vm.install will now use fdisk to create a new disk partition filling the
unallocated space. The root partion is then extended to fill this free
space.