pages: don't render input list until blueprint is loaded
The input list allows users to add and remove components from a
blueprint. These actions can not occur if the blueprint is still being
fetched. By disabling the input list until the blueprint is loaded the
user is prevented from interacting with the list when it may reflect
inaccurate data.