enable filtering with multiple values or wildcards
If multiple filter values are provided separated by comma and/or space,
then these separators (i.e. " " or ", ") are converted to commas (i.e. ",").
If the filter value includes at least one wildcard, then the filter value
provided to the api uses the wildcards as entered, otherwise, the filter
value is wrapped with wildcards.
If filter value happens to start or end with a separator (space and/or
comma) then these are removed.

fixes #1026