test: add rhel-8-beta repo as well
In order to work both against RHEL8.3 Beta, and RHEL8.3 GA,
osbuild-composer had to split the repo files for the two cases. This was
done in commit
https://github.com/osbuild/osbuild-composer/commit/cc677dea3cabcf44c5debd87c5dbd99f2a3ad8c3
which will have to be back-ported to RHEL8.3. Make sure the
cockpit-composer tests continue to work when that's been done.

Cherry-picked from master commit 14208ae11