wizard: translate all strings
The "Blueprint" string and the sidebar label strings were not marked for
translation. Also, the "Type" string was marked for translation but
using a different format than the rest of the project. These are now
fixed.

The step names used in the sidebar are also used when controlling
progress through the wizard. The active step name is now pulled from the
activeStep's props where the defaultMessage of the translatable string
is located.

All strings in the wizard are now marked for translation.