pages: fix blueprints page error state
When a user enters the blueprints page and osbuild-composer is not
running they are prompted to start the socket. This same error state
displays if the user is not root but the button to start the socket
is disabled and alerts the user they don't have permissions. The
check-service test is updated to accept the log messages associated with
trying to access the api without root privileges.