components: remove shortid
The short id package is only used for a key value and is not needed.
Instead the component name is now used for the key. The component names
will always be unique so they will work as a key.

Cherry-picked from master commit f51575f012d1e75fdbcfcffebb87bdb9512c0029