package.json: upgrade sass-loader to be compatible with node 14
This change fixes the test_rpmbuild test. This test was failing after
fedora 33 became the latest version supported by docker. Our docker
image uses fedora:latest and fedora 33 has node 14. The sass-loader
version we were using has a dependency on a version of node-sass that
does not support node 14. Upgrading sass-loader fixes this. Also, the
sass package is added as a dev dependency.