test: disable worker socket on all distros
We do not test against/release in any distros that do not contain
osbuild-composer v25 which includes support for local workers. In order
to completely disable osbuild-composer the worker socket needs to be
disabled alongside the osbuild-composer service and weldr api socket.
These are now disabled on all distros not just fedora 32 and 33.