wizard: add URL parameter to Edge commit
This parameter allows setting a URL to an associated OSTree repository.

This makes the parent commit optional. Change the help text of the
parent field to reflect that.

Bump required osbuild-composer version to the one that gained the `url`
feature.

Based on a patch by Jacob Kozol <jacobdkozol@gmail.com>