CreateImageUpload: add support for new ostree image types
The RHEL for Edge Container and Installer image types are now supported.

The RHEL for Edge Container (.tar) image type supports the optional
ostree fields for a parent commit, a ref for the name of the branch,
and a url for the parent repository.

The RHEL for Edge Installer (.iso) image type supports the required
ostree fields of a ref for the name of the branch and a url for the
parent repository.