test: simplify testUserAccount
When testing the user account modal and list there is no need to test
composer-cli. The goals of these tests are to test the UI against
osbuild-composer. Calling composer-cli is unneccesary since it should
not be treated as more of a source of truth than the UI. Therefore, the
testUserAccount test is simplified.