workflows: split up and rewrite po-refresh
Take the two jobs from the current po-refresh workflow and split them
into two separate workflows with more descriptive names:

- weblate-sync-pot
- weblate-sync-po

The former now runs daily; this avoids lag for translators and has no
extra cost. The latter continues to run weekly, as that causes a PR
which needs to be reviewed, and it's generally not necessary to update
them more often.

Rework the weblate-sync-po workflow:

- we now manually clone the weblate repo and copy the files over for
ourselves.

- we use the GitHub Actions token (appropriately constrained) to push to
our repository and open the PR.

This gets rid of the last user of the `COCKPITUOUS_TOKEN` secret.