test: add check-wizard test
The basic image build tests from check-package and check-image are rewritten
to fit the new DDF Wizard. Instead of openstack we will build a QEMU  qcow2
as the basic image type. It is customized and built. Then the success of
the build is checked alongside downloading and deleting the image.