Wizard: Fix packages step bug
Fixes #1468. The Packages step uses a useState() hook to manage the
state of the selected packages. The state is updated by a useEffect()
hook. The bug occured because the useEffect() hook loaded the package
list from the blueprint. However, the package list in the form state
will be different than the package list in the blueprint if the user
adds or removes packages. Therefore, the useEffect() hook now first
checks whether the form state has a "selected-packages" field. If it
does, then the form state's package list is used to set the selected
packages. If it does not (which is the case the first time the Packages
step is opened) the package list is loaded from the blueprint.