Wizard: Fix number of dependencies shown on Review step
This commit fixes a bug where the number of dependencies displayed on
the review step is incorrect.

At one point, the number of dependencies were determined by depsolving
the blueprint. This was sensible because at that time adding or removing
packages resulted in a commit to the blueprint workspace.

However, currently, adding and removing packages does not cause a commit
to the blueprint workspace. Therefore, the number of dependencies
calculated as a result of depsolving the blueprint is not necessarily
correct (it will be incorrect if the package list in the form state is
different than the blueprint's package list).

The bug has been fixed by calling composer.getComponentDependencies()
instead of composer.depsolveBlueprint().

It should be noted, however, that this changes the calculated number of
dependencies and the behavior is divergent from the original Edit
Blueprint page's (which has been removed from the unified branch) behavior.

Previously, the number of dependencies was the total number of packages
to be installed on the system minus the number of selected packages and
modules. This number included the dependencies that would be installed
for the base image but were not dependencies for the selected packages.

The number of dependencies displayed is the size of a set of the
dependencies of the selected packages, and will most likely not be equal
to the total number of packages actually installed since the base
packages are not necessarily included. A set is used because packages
may share dependencies.