Wizard: Package step improvement and bugfix
Previously, only the blueprint name was available in the form state.
Now, the full blueprint is available in the form state - and includes
the list of selected packages. This allows us to greatly simplify the
useEffect() hook that is used to set the chosenPackages state, as the
list of chosen packages can simply be read directly from the blueprint.
Previously the blueprint had to be depsolved.

Note that an async api call is still required to fetch the package
summaries.

This commit also fixes 1494. This commit fixes a bug where if a package
was searched for in `Available packages` that was in the list of
selected packages, it would incorrectly appear in the `Available packages`
list if available package list was fetched before the chosenPackages
state was set. This was because before setting availablePackages, the
search results are filtered against the chosenPackages, which may not
have loaded yet.

To fix this, when the search results from the available packages are
returned, they are now filtered against the list of selected packages
from the form state rather than the chosenPackages state from the
component.