refactor webpack, router, babel, and translations
The webpack config has been simplified and modernized. Unneeded imports
are cleared from it. The folder public/ has now been refactored to dist/
and the dist build process is simpler. ejs is no longer used in favor of
an index.html file. Our babel config is modernized and moved to a
.babelrc.json.

The routing is now handled by react-router. This simplifies the routing
greatly and gives us quicker flexibility to modify routing and props
passing.

The translations workflow is improved to work on a translations/ dir
with translations imported from weblate directly. The npm scripts are
also simplified to reduce usage of unnneccesray library usage to convert
pot files.