Wizard: split image creation and blueprint creation
The blueprint is created/editted with all customizations specific to the
blueprint and blueprint.customizations. This Wizard contains a
stateToCustomizations function which includes handling for
customizations not yet supported in the UI but that are supported by the
weldr api. This is to simplify adding customizations going forward.

The CreateImageWizard is now refactored to only include fields required
to create an image. The composeRequest object has the blueprint name,
image type, image size, ostree, and upload fields. Therefore, these are
the only fields specified in the Wizard.